requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx==0.25.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL driver

# Web scraping and processing
beautifulsoup4==4.12.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-multipart==0.0.6

# AI Integration
groq==0.4.0